+++ C++/kanade-algorithm.cpp
